rename `HasQuadraticProgram` to `QuadraticProgramElement` (2)
